Add euclidean calculation
Added calculation for euclidean movement rule. Decimal numbers will be displayed and difficult terrain will be added as one grid cell irregardless of your movement direction. This way it's easier to understand during play, as it's not "multiply the movement cost" but "additional movement cost" for difficult terrain.

Sorry for crude code formating. There might be a better way to implement this feature nevertheless. For me it was just a quick-and-dirty fix of the problem.